- MyApp => Appに変更、フォルダを新たに作成してmain.dartから分離しapp.dartに引っ越し - lib配下にある(value)を(_)に変更
